requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
boto3==1.36.6
botocore==1.36.6
cachetools==5.3.1
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
flatbuffers==25.1.24
fonttools==4.55.6
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.12.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
jmespath==1.0.1
joblib==1.3.1
keras==3.8.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.0.post2
llvmlite==0.40.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.0.5
namex==0.0.8
networkx==3.1
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
optree==0.14.0
packaging==23.1
pillow==11.1.0
pooch==1.6.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.19.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
tensorboard==2.17.1
tensorboard-data-server==0.7.1
tensorflow==2.17.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.18
typing_extensions==4.5.0
urllib3==2.3.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
